job-agent/requirements.txt
playwright>=1.42.0
anthropic>=0.25.0
python-docx>=1.1.0
streamlit>=1.32.0
pandas>=2.2.0
altair>=5.3.0
python-dotenv>=1.0.0
rich>=13.7.0
beautifulsoup4>=4.12.0
httpx>=0.27.0
apscheduler>=3.10.0
